Refactored smoke tests configurations
