+++ .github/workflows/docker.yml
@@ -28,21 +28,13 @@ jobs:
     - name: Build Docker container
       run: make docker-build
 
-  smoke-tests_celery4:
-    runs-on: ubuntu-latest
-    timeout-minutes: 10
-    steps:
-    - uses: actions/checkout@v4
-    - name: "Build smoke tests container: pypi - celery 4"
-      run: docker build -f t/smoke/workers/pypi --build-arg CELERY_VERSION="4" .
-
   smoke-tests_dev:
     runs-on: ubuntu-latest
     timeout-minutes: 10
     steps:
     - uses: actions/checkout@v4
     - name: "Build smoke tests container: dev"
-      run: docker build -f t/smoke/workers/dev .
+      run: docker build -f t/smoke/workers/docker/dev .
 
   smoke-tests_latest:
     runs-on: ubuntu-latest
@@ -50,7 +42,7 @@ jobs:
     steps:
     - uses: actions/checkout@v4
     - name: "Build smoke tests container: latest"
-      run: docker build -f t/smoke/workers/pypi .
+      run: docker build -f t/smoke/workers/docker/pypi .
 
   smoke-tests_pypi:
     runs-on: ubuntu-latest
@@ -58,4 +50,12 @@ jobs:
     steps:
     - uses: actions/checkout@v4
     - name: "Build smoke tests container: pypi"
-      run: docker build -f t/smoke/workers/pypi --build-arg CELERY_VERSION="5.3" .
+      run: docker build -f t/smoke/workers/docker/pypi --build-arg CELERY_VERSION="5" .
+
+  smoke-tests_legacy:
+    runs-on: ubuntu-latest
+    timeout-minutes: 10
+    steps:
+    - uses: actions/checkout@v4
+    - name: "Build smoke tests container: legacy"
+      run: docker build -f t/smoke/workers/docker/pypi --build-arg CELERY_VERSION="4" .
